chatify-api/requirements.txt
Flask==3.1.0
flask-cors
pymongo==4.11.1
PyJWT
bcrypt
redis
